[doc] Add anchor links to headers in api_doc files
Currently, there is no good way to link to sections of the API
documentation. While there are links to sections/methods in the sidebar
and in 'table of contents'-like areas of the doc, it would be better if
there was a way to grab a link directly from the section header/method
name itself.

This commit achieves this by simply adding a markdown link to headers
that include an #id in braces. For example:
    ### This Awesome Header {#awesome-header}
Becomes
    ### [This Awesome Header](#awesome-header) {#awesome-header}

The hope is that by making it easier to reference specific sections of
the documentation, it will encourage third-parties to create tutorials
and pages that point directly to the subject matter at hand.
Additionally, it will become easier for colleagues to share reference
links to one another.

This was done using the script available here:
https://github.com/samjabrahams/tensorflow_util/blob/master/py/add_header_permalink.py

It's possible that making the entire header a link is not the best way
to set up anchor links, but changing the specific implementation is not
a big deal. For example, if it's decided that there should be a 'chain'
icon to the side of headers instead of making the entire header a link,
that should be easy to patch in.

One thing that should be considered, if this patch is accepted as-is,
is the CSS styling of headers vs links on the TensorFlow website. If
things go wonky (i.e. headers become orange), I believe the following
addition to main.css will suffice:
```
.block-copy h1 a, .block-copy h2 a, .block-copy h3 a, .block-copy h4 a,
.block-copy h5 a {
    color: #000;
}
```